:lipstick: Redesign interface of Thorax assistant
The Thorax Setup Assistant has been redesigned to feel more sleek and
intuitive. The main window now takes a smaller portion of the screen,
and the background for the release is displayed behind the setup window
with a dynamic blur effect at runtime. The text field's color has been
changed to be more readable, and it is positioned higher on the screen.
The title and detail text sizes have shrunk to make it more legible.

Additionally, an issue where MISSING_OOBE_SERVICE would crash the system
has been resolved by writing a more stringent check on how the setup
assistant is detected. This will need to be improved on in the future,
preferably with a permission in the manifest.

Finally, a bug in CAUserData._create_user_file where the username was
not being properly stripped from whitespaces has been resolved by
fixing which username was being written to in the user file.